--- a/proto/binlogdata.proto
+++ b/proto/binlogdata.proto
@@ -19,6 +19,7 @@ limitations under the License.
 // replication only.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/binlogdata";
 
 package binlogdata;