--- a/proto/binlogservice.proto
+++ b/proto/binlogservice.proto
@@ -19,6 +19,7 @@ limitations under the License.
 // filtered replication only.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/binlogservice";
 
 package binlogservice;