--- a/proto/tabletmanagerdata.proto
+++ b/proto/tabletmanagerdata.proto
@@ -18,6 +18,7 @@ limitations under the License.
 // RPC calls to VtTablet for the management API.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/tabletmanagerdata";
 
 package tabletmanagerdata;