--- a/proto/tabletmanagerservice.proto
+++ b/proto/tabletmanagerservice.proto
@@ -18,6 +18,7 @@ limitations under the License.
 // calls to VtTablet.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/tabletmanagerservice";
 
 package tabletmanagerservice;