--- a/proto/throttlerdata.proto
+++ b/proto/throttlerdata.proto
@@ -17,6 +17,7 @@ limitations under the License.
 // Data structures for the throttler RPC interface.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/throttlerdata";
 
 package throttlerdata;