--- a/proto/vtctldata.proto
+++ b/proto/vtctldata.proto
@@ -18,6 +18,7 @@ limitations under the License.
 // you to use vtctld as a server for vt commands.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/vtctldata";
 
 package vtctldata;