--- a/proto/vtgate.proto
+++ b/proto/vtgate.proto
@@ -17,6 +17,7 @@ limitations under the License.
 // Data definitions for service vtgateservice.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/vtgate";
 
 package vtgate;