--- a/proto/vtgateservice.proto
+++ b/proto/vtgateservice.proto
@@ -18,6 +18,7 @@ limitations under the License.
 // This is the main entry point to Vitess.
 
 syntax = "proto3";
+option go_package = "vitess.io/vitess/go/vt/proto/vtgateservice";
 
 option java_package="io.vitess.proto.grpc";