--- a/py/vtproto/vtworkerdata_pb2.py
+++ b/py/vtproto/vtworkerdata_pb2.py
@@ -20,7 +20,7 @@ DESCRIPTOR = _descriptor.FileDescriptor(
   name='vtworkerdata.proto',
   package='vtworkerdata',
   syntax='proto3',
-  serialized_pb=_b('\n\x12vtworkerdata.proto\x12\x0cvtworkerdata\x1a\rlogutil.proto\"-\n\x1d\x45xecuteVtworkerCommandRequest\x12\x0c\n\x04\x61rgs\x18\x01 \x03(\t\"?\n\x1e\x45xecuteVtworkerCommandResponse\x12\x1d\n\x05\x65vent\x18\x01 \x01(\x0b\x32\x0e.logutil.Eventb\x06proto3')
+  serialized_pb=_b('\n\x12vtworkerdata.proto\x12\x0cvtworkerdata\x1a\rlogutil.proto\"-\n\x1d\x45xecuteVtworkerCommandRequest\x12\x0c\n\x04\x61rgs\x18\x01 \x03(\t\"?\n\x1e\x45xecuteVtworkerCommandResponse\x12\x1d\n\x05\x65vent\x18\x01 \x01(\x0b\x32\x0e.logutil.EventB+Z)vitess.io/vitess/go/vt/proto/vtworkerdatab\x06proto3')
   ,
   dependencies=[logutil__pb2.DESCRIPTOR,])
 
@@ -108,4 +108,6 @@ ExecuteVtworkerCommandResponse = _reflection.GeneratedProtocolMessageType('Execu
 _sym_db.RegisterMessage(ExecuteVtworkerCommandResponse)
 
 
+DESCRIPTOR.has_options = True
+DESCRIPTOR._options = _descriptor._ParseOptions(descriptor_pb2.FileOptions(), _b('Z)vitess.io/vitess/go/vt/proto/vtworkerdata'))
 # @@protoc_insertion_point(module_scope)