@dweitzman 2018-06-08
This seems to work for me on OS X and linux, but it's probably a good idea to try running bootstrap & make proto yourself to double-check that it works on your systems as well