@dweitzman 2018-06-08
Related: https://github.com/vitessio/vitess/issues/3566