@dweitzman 2018-06-08
It seems like the travis / bootstrap image / non-reproducible build situatino could make this kind of change awkward.

This sequence of commands seems to work fine for me locally:

`./docker/bootstrap/build.sh percona && make docker_base_percona && docker run -ti vitess/base:percona make unit_test`