@sougou 2018-06-12
I've confirmed that the build will (likely) succeed once we push the updated docker images.
However, when I ran `make proto`, the python files produced a diff. Can you check on that?

I also did a quick benchmark, and I did not see any change in performance. This is surprising. The numbers were so identical that I'm doubting myself. Maybe someone else could run their own and share results?