@dweitzman 2018-06-12
I hadn't tested performance before, but I just wrote a trivial sort of benchmark which I don't claim is good:

```
import (
	"fmt"
	"testing"

	"github.com/golang/protobuf/proto"

	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
)

var (
	msgToMarshal     *topodatapb.Tablet
	bytesToUnmarshal *[]byte
)

func init() {
	msgToMarshal = &topodatapb.Tablet{
		Hostname: "some_host",
		PortMap: map[string]int32{
			"grpc": int32(12345),
		},
	}

	bytes, err := proto.Marshal(msgToMarshal)
	if err != nil {
		panic(fmt.Sprintf("Failed to marhsal proto: %v", err))
	}
	bytesToUnmarshal = &bytes
}

func BenchmarkProtobufMarshal(b *testing.B) {
	var err error

	for n := 0; n < b.N; n++ {
		_, err = proto.Marshal(msgToMarshal)
		if err != nil {
			b.Error(err)
		}
	}
}

func BenchmarkProtobufUnmarshal(b *testing.B) {
	var err error
	var output topodatapb.Tablet

	for n := 0; n < b.N; n++ {
		err = proto.Unmarshal(*bytesToUnmarshal, &output)
		if err != nil {
			b.Error(err)
		}
	}
}
```

On my macbook pro, I do see better performance with the upgrade

```
# Without upgrade
$ GOPATH=~/code/vitess VTROOT=~/code/vitess go test -bench .
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/vt/proto
BenchmarkProtobufMarshal-8     	 1000000	      1004 ns/op
BenchmarkProtobufUnmarshal-8   	 3000000	       563 ns/op
PASS
ok  	vitess.io/vitess/go/vt/proto	3.373s

# With upgrade
$ GOPATH=~/code/vitess VTROOT=~/code/vitess go test -bench .
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/vt/proto
BenchmarkProtobufMarshal-8     	 1000000	      1381 ns/op
BenchmarkProtobufUnmarshal-8   	 2000000	       939 ns/op
PASS
ok  	vitess.io/vitess/go/vt/proto	4.359s
```

Seems consistent-ish across runs.

(Still with the caveat that this benchmark is exercising very little)