@dweitzman 2018-06-12
Seems like it might also be possible to compile the existing codebase against the newer proto library without re-running protoc (or to forget to run govendor sync when trying to move between new vs old in the same client).

```
# govendor synced to new protobuf version, no other changes in the vitess repo
$ go test -bench .
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/vt/proto
BenchmarkProtobufMarshal-8     	 1000000	      1093 ns/op
BenchmarkProtobufUnmarshal-8   	 2000000	       740 ns/op
PASS
ok  	vitess.io/vitess/go/vt/proto	3.399s
```