@sougou 2018-06-29
When I pulled this and ran make proto, it regenerated the python files. Can you check if the same thing happens for you? If so, can you check those changes in?