@dweitzman 2018-07-10
Hmm. I tried running "rm py/vtproto/*.proto && make proto" and it generates the exact same python files for me (except py/vtproto/__init__.py is still missing -- maybe it's hand-written?)